Added asset connection setup
