+++ egeria-connector-jdbc/src/main/java/org/odpi/openmetadata/adapters/connectors/integration/jdbc/transfer/JdbcMetadataTransfer.java
@@ -376,137 +376,9 @@ public class JdbcMetadataTransfer {
     }
 
     private void createAssetConnection(DatabaseElement databaseElement){
-        String methodName = "createAssetConnection";
-
-        String connectorTypeQualifiedName = jdbcMetadata.getConnectorTypeQualifiedName();
-        if(StringUtils.isBlank(connectorTypeQualifiedName)){
-            auditLog.logMessage("Missing connector type qualified name. Skipping asset connection setup",
-                    null);
-            return;
-        }
-        String connectorTypeGuid = determineConnectorTypeGuid(connectorTypeQualifiedName);
-        if(StringUtils.isBlank(connectorTypeGuid)){
-            auditLog.logMessage("Missing connector type guid. Skipping asset connection setup",
-                    null);
-            return;
-        }
-
-        String databaseGuid = databaseElement.getElementHeader().getGUID();
-        if(StringUtils.isBlank(databaseGuid)){
-            auditLog.logMessage("Missing database guid. Skipping asset connection setup",
-                    null);
-            return;
-        }
-
-        ConnectionProperties connectionProperties = createConnectionProperties(databaseElement);
-        String connectionGuid = determineConnectionGuid(connectionProperties);
-        if(StringUtils.isBlank(connectionGuid)){
-            auditLog.logMessage("Missing connection guid. Skipping asset connection setup",
-                    null);
-            return;
-        }
-
-        EndpointProperties endpointProperties = createEndpointProperties(connectionProperties);
-        String endpointGuid = determineEndpointGuid(endpointProperties);
-        if(StringUtils.isBlank(endpointGuid)){
-            auditLog.logMessage("Missing endpoint guid. Skipping asset connection setup",
-                    null);
-            return;
-        }
-
-        try {
-            databaseIntegratorContext.setupConnectorType(connectionGuid, connectorTypeGuid);
-            databaseIntegratorContext.setupAssetConnection(databaseGuid,
-                    databaseElement.getDatabaseProperties().getDescription(), connectionGuid);
-            databaseIntegratorContext.setupEndpoint(connectionGuid, endpointGuid);
-        } catch (InvalidParameterException | UserNotAuthorizedException | PropertyServerException e) {
-            auditLog.logMessage("Setting up connection (guid: " + connectionGuid
-                            + "), connector type (guid: " + connectorTypeQualifiedName
-                            + ") and asset (guid: " + databaseGuid + " . Ignoring",
-                    ERROR_UPSERTING_INTO_OMAS.getMessageDefinition(methodName, e.getMessage()));
-        }
-    }
-
-    private ConnectionProperties createConnectionProperties(DatabaseElement databaseElement){
-        ConnectionProperties connectionProperties = new ConnectionProperties();
-        connectionProperties.setDisplayName(databaseElement.getDatabaseProperties().getDisplayName() + " Connection");
-        connectionProperties.setQualifiedName(databaseElement.getDatabaseProperties().getQualifiedName() + "::connection");
-        connectionProperties.setConfigurationProperties(databaseElement.getDatabaseProperties().getExtendedProperties());
-
-        return connectionProperties;
-    }
-
-    private String determineConnectionGuid(ConnectionProperties connectionProperties){
-        String methodName = "determineConnectionGuid";
-        try {
-            Optional<List<ConnectionElement>> connections = Optional.ofNullable(
-                    databaseIntegratorContext.getConnectionsByName(connectionProperties.getQualifiedName(),
-                            0, 0));
-            if(connections.isPresent()){
-                if(connections.get().size() == 1){
-                    return connections.get().get(0).getElementHeader().getGUID();
-                }
-            }else{
-                return databaseIntegratorContext.createConnection(connectionProperties);
-            }
-        } catch (UserNotAuthorizedException | InvalidParameterException | PropertyServerException e) {
-            auditLog.logMessage("Determining connection guid",
-                    ERROR_WHEN_SETTING_ASSET_CONNECTION.getMessageDefinition(methodName));
-        }
-
-        return null;
-    }
-
-    private String determineConnectorTypeGuid(String connectorTypeQualifiedName){
-        String methodName = "determineConnectorTypeGuid";
-        try{
-            Optional<List<ConnectorTypeElement>> connectorTypes = Optional.ofNullable(
-                    databaseIntegratorContext.getConnectorTypesByName(connectorTypeQualifiedName, 0, 0));
-            if(connectorTypes.isPresent()){
-                if(connectorTypes.get().size() == 1){
-                    return connectorTypes.get().get(0).getElementHeader().getGUID();
-                }
-            }
-        } catch (UserNotAuthorizedException | InvalidParameterException | PropertyServerException e) {
-            auditLog.logMessage("Determining connector type guid",
-                    ERROR_WHEN_SETTING_ASSET_CONNECTION.getMessageDefinition(methodName));
-        }
-        return null;
-    }
-
-    private EndpointProperties createEndpointProperties(ConnectionProperties connectionProperties){
-        String methodName = "createEndpointProperties";
-
-        EndpointProperties endpointProperties = new EndpointProperties();
-        endpointProperties.setDisplayName(connectionProperties.getDisplayName() + " Endpoint");
-        endpointProperties.setQualifiedName(connectionProperties.getQualifiedName()+"::endpoint");
-        try {
-            endpointProperties.setAddress(jdbcMetadata.getUrl());
-        } catch (SQLException sqlException) {
-            auditLog.logMessage("Reading url from jdbc metadata",
-                    ERROR_READING_JDBC.getMessageDefinition(methodName, sqlException.getMessage()));
-        }
-
-        return endpointProperties;
-    }
-    
-    private String determineEndpointGuid(EndpointProperties endpointProperties){
-        String methodName = "determineEndpointGuid";
-        try{
-            Optional<List<EndpointElement>> endpoints = Optional.ofNullable(
-                    databaseIntegratorContext.findEndpoints(endpointProperties.getQualifiedName(), 0, 0));
-            if(endpoints.isPresent()){
-                if(endpoints.get().size() == 1) {
-                    return endpoints.get().get(0).getElementHeader().getGUID();
-                }
-            }else{
-                return databaseIntegratorContext.createEndpoint(endpointProperties);
-            }
-        } catch (UserNotAuthorizedException | InvalidParameterException | PropertyServerException e) {
-            auditLog.logMessage("Determining endpoint guid",
-                    ERROR_WHEN_SETTING_ASSET_CONNECTION.getMessageDefinition(methodName));
-        }
-        return null;
+        DatabaseConnectionConsumer databaseConnectionConsumer =
+                new DatabaseConnectionConsumer(databaseIntegratorContext, auditLog, jdbcMetadata);
+        databaseConnectionConsumer.accept(databaseElement);
     }
 
 }
